Clean up source files
- Split MemberPolicies.cs to conform to 
  the 1 file per class layout. The classes
  were too big to warrant an exemption
  and navigating the file has been
  a constant source of annoyance.

- Moved the one remaining tenant of MemberEnumerator.cs
  into Shared.cs as it had nothing to do with
  enumerating members.
