+++ src/System.Private.Reflection.Core/src/System.Private.Reflection.Core.csproj
@@ -46,12 +46,17 @@
     <Compile Include="System\Reflection\Runtime\Assemblies\RuntimeAssembly.GetTypeCore.CaseInsensitive.cs" />
     <Compile Include="System\Reflection\Runtime\Assemblies\RuntimeAssembly.GetTypeCore.CaseSensitive.cs" />
     <Compile Include="System\Reflection\Runtime\Assemblies\RuntimeAssemblyName.cs" />
-    <Compile Include="System\Reflection\Runtime\BindingFlagSupport\DefaultBinder.cs" />
+    <Compile Include="System\Reflection\Runtime\BindingFlagSupport\ConstructorPolicies.cs" />
     <Compile Include="System\Reflection\Runtime\BindingFlagSupport\DefaultBinder.CanConvert.cs" />
-    <Compile Include="System\Reflection\Runtime\BindingFlagSupport\MemberEnumerator.cs" />
+    <Compile Include="System\Reflection\Runtime\BindingFlagSupport\DefaultBinder.cs" />
+    <Compile Include="System\Reflection\Runtime\BindingFlagSupport\EventPolicies.cs" />
+    <Compile Include="System\Reflection\Runtime\BindingFlagSupport\FieldPolicies.cs" />
     <Compile Include="System\Reflection\Runtime\BindingFlagSupport\MemberPolicies.cs" />
     <Compile Include="System\Reflection\Runtime\BindingFlagSupport\MemberTypeIndex.cs" />
+    <Compile Include="System\Reflection\Runtime\BindingFlagSupport\MethodPolicies.cs" />
     <Compile Include="System\Reflection\Runtime\BindingFlagSupport\NameFilter.cs" />
+    <Compile Include="System\Reflection\Runtime\BindingFlagSupport\NestedTypePolicies.cs" />
+    <Compile Include="System\Reflection\Runtime\BindingFlagSupport\PropertyPolicies.cs" />
     <Compile Include="System\Reflection\Runtime\BindingFlagSupport\QueriedMemberList.cs" />
     <Compile Include="System\Reflection\Runtime\BindingFlagSupport\QueryResult.cs" />
     <Compile Include="System\Reflection\Runtime\BindingFlagSupport\QueryResult.Enumerator.cs" />
